Changed units for pedals and output to 0-1, added OTA

+++ include/settings.h
@@ -31,21 +31,21 @@ class Settings {
         char hostname[32];
         char playstation[32];
 
-        unsigned int   decel_pin1;
-        unsigned int   decel_pin2;
-        unsigned int   decel_output;
-        double         decel_threshold;
-        enum mode      decel_mode;
-        unsigned int   decel_min_pedal;
-        double         decel_min_speed;
+        uint8_t     decel_pin1;
+        uint8_t     decel_pin2;
+        float       decel_output;
+        float       decel_threshold;
+        enum mode   decel_mode;
+        float       decel_min_pedal;
+        float       decel_min_speed;
 
-        unsigned int   accel_pin1;
-        unsigned int   accel_pin2;
-        unsigned int   accel_output;
-        double         accel_threshold;
-        enum mode      accel_mode;
-        unsigned int   accel_min_pedal;
-        double         accel_min_speed;
+        uint8_t     accel_pin1;
+        uint8_t     accel_pin2;
+        float       accel_output;
+        float       accel_threshold;
+        enum mode   accel_mode;
+        float       accel_min_pedal;
+        float       accel_min_speed;
 
         void load(const char* filename);
         void save(const char* filename);
